Correct bidrectional hierarchy to allow context to work.
