+++ pyaestro/abstracts/graphs/__init__.py
@@ -1,9 +1,7 @@
 from pyaestro.abstracts.graphs._graph import Graph
-from pyaestro.abstracts.graphs._interfaces import GraphEdgeInterface
 
 __all__ = (
     "AcyclicGraph",
     "BidirectionalGraphMixin",
     "Graph",
-    "GraphEdgeInterface",
 )
