--- pyaestro/abstracts/graphs/_interfaces.py
@@ -1,40 +1,0 @@
-from __future__ import annotations
-
-from abc import ABC, abstractmethod
-from typing import Hashable
-
-from pyaestro.typing import Comparable
-
-
-class GraphEdgeInterface(ABC):
-    @abstractmethod
-    def add_edge(
-        self, a: Hashable, b: Hashable, weight: Comparable = 0
-    ) -> None:
-        """Add an edge to the graph.
-
-        Args:
-            a (Hashable): Key identifying side 'a' of an edge.
-            b (Hashable): Key identifying side 'b' of an edge.
-            weight(Comparable): Weight of the edge between 'a' and 'b'.
-            Defaults to 0 for unweighted.
-
-        Raises:
-            KeyError: Raised when either node 'a' or node 'b'
-            do not exist in the graph.
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def remove_edge(self, a: Hashable, b: Hashable) -> None:
-        """Remove an edge to the graph.
-
-        Args:
-            a (Hashable): Key identifying side 'a' of an edge.
-            b (Hashable): Key identifying side 'b' of an edge.
-
-        Raises:
-            KeyError: Raised when either node 'a' or node 'b'
-            do not exist in the graph.
-        """
-        raise NotImplementedError
